# PERF: only copy input array in _replace_nan() if there are nans to replace