@eric-wieser 2019-12-02
Does this have any effect on the public API, or is a copy always made from the result of `_replace_nan` before returning it to the user?