@qwhelan 2019-12-03
@eric-wieser Every function in `nanfunctions.py` is a reduction, so a dimension reduction is always done on the output of `_replace_nan()`. Public functions are:
```python
__all__ = [
    'nansum', 'nanmax', 'nanmin', 'nanargmax', 'nanargmin', 'nanmean',
    'nanmedian', 'nanpercentile', 'nanvar', 'nanstd', 'nanprod',
    'nancumsum', 'nancumprod', 'nanquantile'
    ]
```